pyproject.toml
[project]
name = "contract-agent"
version = "0.1.0"
description = "企业合同智能助手 - 基于 AI 的合同管理和审查系统"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["contract", "ai", "llm", "fastapi", "vue"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.16",
    "python-docx>=1.1.0",
    "docxtpl>=0.16.0",
    "litellm>=1.40.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "tomlkit>=0.12.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "json-repair>=0.25.0",
    "tenacity>=8.2.0",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/contract-agent"
Repository = "https://github.com/yourusername/contract-agent"
Documentation = "https://github.com/yourusername/contract-agent#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/src"]

[tool.black]
line-length = 80
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "google"
line_length = 80
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 80
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
